Rename BRS namespaces
