+++ source/B2BApi/Functions/EnqueueMessages/BRS_023_027/Activities/EnqueueWholesaleResultsBaseActivity.cs
@@ -13,49 +13,51 @@
 // limitations under the License.
 
 using System.Diagnostics;
-using Energinet.DataHub.EDI.B2BApi.Functions.EnqueueMessages.BRS_021_023.Model;
-using Energinet.DataHub.EDI.OutgoingMessages.Infrastructure.Databricks.EnergyResults.Queries;
+using Energinet.DataHub.EDI.B2BApi.Functions.EnqueueMessages.BRS_023_027.Model;
+using Energinet.DataHub.EDI.OutgoingMessages.Infrastructure.Databricks.WholesaleResults.Queries;
 using Energinet.DataHub.EDI.OutgoingMessages.Interfaces;
 using Energinet.DataHub.EDI.OutgoingMessages.Interfaces.Models;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 
-namespace Energinet.DataHub.EDI.B2BApi.Functions.EnqueueMessages.BRS_021_023.Activities;
+namespace Energinet.DataHub.EDI.B2BApi.Functions.EnqueueMessages.BRS_023_027.Activities;
 
-public abstract class EnqueueEnergyResultsBaseActivity(
+public abstract class EnqueueWholesaleResultsBaseActivity(
     ILogger logger,
     IServiceScopeFactory serviceScopeFactory,
-    EnergyResultEnumerator energyResultEnumerator)
+    WholesaleResultEnumerator wholesaleResultEnumerator)
 {
     private readonly ILogger _logger = logger;
     private readonly IServiceScopeFactory _serviceScopeFactory = serviceScopeFactory;
-    private readonly EnergyResultEnumerator _energyResultEnumerator = energyResultEnumerator;
+    private readonly WholesaleResultEnumerator _wholesaleResultEnumerator = wholesaleResultEnumerator;
 
-    protected async Task<int> EnqueueEnergyResults<TQueryResult>(EnqueueMessagesInput input, EnergyResultQueryBase<TQueryResult> query)
+    protected async Task<int> EnqueueWholesaleResults<TQueryResult>(EnqueueMessagesForActorInput input, WholesaleResultQueryBase<TQueryResult> query)
         where TQueryResult : OutgoingMessageDto
     {
         var numberOfHandledResults = 0;
         var numberOfFailedResults = 0;
 
         _logger.LogInformation(
-            "Starting enqueuing messages for energy query, type: {QueryType}, calculation id: {CalculationId}, event id: {EventId}",
+            "Starting enqueuing messages for wholesale query, type: {QueryType}, actor: {Actor}, calculation id: {CalculationId}, event id: {EventId}",
             query.GetType().Name,
+            input.Actor,
             input.CalculationId,
             input.EventId);
 
         var activityStopwatch = Stopwatch.StartNew();
         var databricksStopwatch = Stopwatch.StartNew();
-        await foreach (var queryResult in _energyResultEnumerator.GetAsync(query))
+        await foreach (var queryResult in _wholesaleResultEnumerator.GetAsync(query))
         {
             databricksStopwatch.Stop();
             // Only log databricks query time if it took more than 1 second
             if (databricksStopwatch.Elapsed > TimeSpan.FromSeconds(1))
             {
                 _logger.LogInformation(
-                    "Retrieved energy result from databricks, elapsed time: {ElapsedTime}, type: {QueryType}, external id: {ExternalId}, calculation id: {CalculationId}, event id: {EventId}",
+                    "Retrieved energy result from databricks, elapsed time: {ElapsedTime}, type: {QueryType}, external id: {ExternalId}, actor: {Actor}, calculation id: {CalculationId}, event id: {EventId}",
                     databricksStopwatch.Elapsed,
                     query.GetType().Name,
                     queryResult.Result?.ExternalId.Value,
+                    input.Actor,
                     input.CalculationId,
                     input.EventId);
             }
@@ -66,8 +68,7 @@ public abstract class EnqueueEnergyResultsBaseActivity(
                 try
                 {
                     var scopedOutgoingMessagesClient = scope.ServiceProvider.GetRequiredService<IOutgoingMessagesClient>();
-
-                    await EnqueueAndCommitEnergyResult(scopedOutgoingMessagesClient, queryResult.Result!).ConfigureAwait(false);
+                    await EnqueueAndCommitWholesaleResult(scopedOutgoingMessagesClient, queryResult.Result!).ConfigureAwait(false);
 
                     numberOfHandledResults++;
                 }
@@ -76,9 +77,10 @@ public abstract class EnqueueEnergyResultsBaseActivity(
                     numberOfFailedResults++;
                     _logger.LogWarning(
                         ex,
-                        "Enqueue and commit threw exception for energy result, query type: {QueryType}, external id: {ExternalId}, calculation id: {CalculationId}, event id: {EventId}",
+                        "Enqueue and commit threw exception for wholesale result, query type: {QueryType}, external id: {ExternalId}, actor: {Actor}, calculation id: {CalculationId}, event id: {EventId}",
                         query.GetType().Name,
                         queryResult.Result?.ExternalId.Value,
+                        input.Actor,
                         input.CalculationId,
                         input.EventId);
                 }
@@ -92,19 +94,20 @@ public abstract class EnqueueEnergyResultsBaseActivity(
         }
 
         _logger.LogInformation(
-            "Finished enqueuing messages for energy query, elapsed time: {ElapsedTime}, successful results: {SuccessfulResultsCount}, failed results: {FailedResultsCount}, type: {QueryType}, calculation id: {CalculationId}, event id: {EventId}",
+            "Finished enqueuing messages for wholesale query, elapsed time: {ElapsedTime}, successful results: {SuccessfulResultsCount}, failed results: {FailedResultsCount}, type: {QueryType}, actor: {Actor}, calculation id: {CalculationId}, event id: {EventId}",
             activityStopwatch.Elapsed,
             numberOfHandledResults,
             numberOfFailedResults,
             query.GetType().Name,
+            input.Actor,
             input.CalculationId,
             input.EventId);
 
         return numberOfFailedResults > 0
-            ? throw new Exception($"Enqueue messages activity failed. CalculationId='{input.CalculationId}' EventId='{input.EventId}' NumberOfFailedResults='{numberOfFailedResults}' NumberOfHandledResults='{numberOfHandledResults}'")
+            ? throw new Exception($"Enqueue messages activity failed. Actor='{input.Actor}' CalculationId='{input.CalculationId}' EventId='{input.EventId}' NumberOfFailedResults='{numberOfFailedResults}' NumberOfHandledResults='{numberOfHandledResults}'")
             : numberOfHandledResults;
     }
 
-    protected abstract Task EnqueueAndCommitEnergyResult<T>(IOutgoingMessagesClient outgoingMessagesClient, T queryResult)
+    protected abstract Task EnqueueAndCommitWholesaleResult<T>(IOutgoingMessagesClient outgoingMessagesClient, T queryResult)
         where T : OutgoingMessageDto;
 }
