+++ source/B2BApi/Functions/EnqueueMessages/BRS_023_027/Model/EnqueueMessagesInput.cs
@@ -12,13 +12,21 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.EDI.B2BApi.Functions.EnqueueMessages.BRS_021_023.Model;
+using System.Collections.Immutable;
+using Energinet.DataHub.EDI.BuildingBlocks.Domain.Models;
+
+namespace Energinet.DataHub.EDI.B2BApi.Functions.EnqueueMessages.BRS_023_027.Model;
 
 /// <summary>
-/// An immutable input to start the send messages enqueued activity.
+/// An immutable input to start the enqueue messages activity.
 /// </summary>
-public sealed record SendMessagesEnqueuedInput(
-    string OrchestrationInstanceId,
-    string CalculationOrchestrationInstanceId,
+public sealed record EnqueueMessagesInput(
     Guid CalculationId,
-    bool Success);
+    Guid EventId,
+    ImmutableDictionary<string, ActorNumber> GridAreaOwners);
+
+public sealed record EnqueueMessagesForActorInput(
+    Guid CalculationId,
+    Guid EventId,
+    ImmutableDictionary<string, ActorNumber> GridAreaOwners,
+    string Actor);
