+++ source/B2BApi/Functions/EnqueueMessages/BRS_023_027/Model/SendMessagesEnqueuedInput.cs
@@ -12,21 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System.Collections.Immutable;
-using Energinet.DataHub.EDI.BuildingBlocks.Domain.Models;
-
-namespace Energinet.DataHub.EDI.B2BApi.Functions.EnqueueMessages.BRS_021_023.Model;
+namespace Energinet.DataHub.EDI.B2BApi.Functions.EnqueueMessages.BRS_023_027.Model;
 
 /// <summary>
-/// An immutable input to start the enqueue messages activity.
+/// An immutable input to start the send messages enqueued activity.
 /// </summary>
-public sealed record EnqueueMessagesInput(
+public sealed record SendMessagesEnqueuedInput(
+    string OrchestrationInstanceId,
+    string CalculationOrchestrationInstanceId,
     Guid CalculationId,
-    Guid EventId,
-    ImmutableDictionary<string, ActorNumber> GridAreaOwners);
-
-public sealed record EnqueueMessagesForActorInput(
-    Guid CalculationId,
-    Guid EventId,
-    ImmutableDictionary<string, ActorNumber> GridAreaOwners,
-    string Actor);
+    bool Success);
